decimal value set for expense also

+++ migrations/20250729183215-create_ledger_income.js
@@ -10,7 +10,7 @@ module.exports = {
         autoIncrement: true,
       },
       amount: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(20, 5),
         allowNull: false,
       },
       description: {
